+++ pennylane/__init__.py
@@ -44,7 +44,7 @@ from .measure import expval, var, sample, probs
 from .ops import *
 from .optimize import *
 from .qnodes import qnode, QNode, QuantumFunctionError
-from .utils import inv, SneakPeak
+from .utils import inv, SneakPeek
 from ._version import __version__
 from .io import *